+++ plugin/web/extensions/sidebar/modals/datacatalog/components/content/DatasetsPage/DatasetTree/File.tsx
@@ -13,7 +13,7 @@ export type Props = {
   onDatasetAdd: (dataset: DataCatalogItem) => void;
   onOpenDetails?: (item?: DataCatalogItem) => void;
   onSelect?: (dataID: string) => void;
-  setSelectedKey: (key: string) => void;
+  setExpandedKeys: React.Dispatch<React.SetStateAction<string[]>>;
 };
 
 const File: React.FC<Props> = ({
@@ -26,7 +26,7 @@ const File: React.FC<Props> = ({
   onDatasetAdd,
   onOpenDetails,
   onSelect,
-  setSelectedKey,
+  setExpandedKeys,
 }) => {
   const handleClick = useCallback(() => {
     onDatasetAdd(item);
@@ -42,9 +42,33 @@ const File: React.FC<Props> = ({
     [selectedID, item],
   );
 
+  const expandAllParentKeys = useCallback(
+    (key: string) => {
+      const keyArr = key.split("-");
+      while (keyArr.length > 1) {
+        keyArr.pop();
+        const parent = keyArr.join("-");
+        setExpandedKeys((prevState: string[]) => {
+          const newExpandedKeys = [...prevState];
+          if (!prevState.includes(parent)) newExpandedKeys.push(parent);
+          return newExpandedKeys;
+        });
+      }
+    },
+    [setExpandedKeys],
+  );
+
   useEffect(() => {
-    if (selected) setSelectedKey(nodeKey);
-  }, [nodeKey, selected, setSelectedKey]);
+    const { selectedDataset } = window as any;
+    if (selectedDataset) {
+      onOpenDetails?.(selectedDataset);
+      onSelect?.(selectedDataset.dataID);
+      if (selected) expandAllParentKeys(nodeKey);
+      setTimeout(() => {
+        (window as any).selectedDataset = undefined;
+      }, 500);
+    }
+  }, [expandAllParentKeys, nodeKey, onOpenDetails, onSelect, selected]);
 
   return (
     <Wrapper nestLevel={nestLevel} selected={selected}>
